DDP-8721 fix osteo release pdfs
