+++ pepper-apis/studybuilder-cli/studies/osteo/patches/osteo-release-v2-pdfs.conf
@@ -1,0 +1,199 @@
+{
+  "pdfs": [
+    {
+      "name": "osproject-release",
+      "filename": "osproject-release",
+      "displayName": "Osteo release pdf",
+      "versions": [
+        {
+          "tag": "v2",
+          "start": null,
+          "end": null,
+          "sources": [
+            {
+              "type": "PARTICIPANT"
+            },
+            {
+              "type": "ACTIVITY",
+              "activityCode": "RELEASE_SELF",
+              "versionTag": "v2"
+            },
+            {
+              "type": "ACTIVITY",
+              "activityCode": "CONSENT",
+              "versionTag": "v2"
+            }
+          ],
+          "files": [
+            {
+              "filepath": "pdfs/release_contact_v2.pdf",
+              "type": "CUSTOM",
+              "substitutions": []
+            },
+            {
+              "filepath": "pdfs/release_physician_v2.pdf",
+              "type": "PHYSICIAN",
+              "fields": {
+                "name": "physician_name",
+                "institution": "physician_institution",
+                "city": "physician_city",
+                "state": "physician_state"
+              }
+            },
+            {
+              "filepath": "pdfs/release_agreement_v2.pdf",
+              "type": "CUSTOM",
+              "substitutions": [
+                {
+                  "type": "TEXT",
+                  "field": "first_name",
+                  "activityCode": "CONSENT",
+                  "questionStableId": "CONSENT_FIRSTNAME"
+                },
+                {
+                  "type": "TEXT",
+                  "field": "second_name",
+                  "activityCode": "CONSENT",
+                  "questionStableId": "CONSENT_LASTNAME"
+                },
+                {
+                  "type": "ACTIVITY_DATE",
+                  "field": "date",
+                  "activityCode": "RELEASE_SELF"
+                }
+              ]
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "name": "osproject-release-parental",
+      "filename": "osproject-release-parental",
+      "displayName": "Osteo parental release pdf",
+      "versions": [
+        {
+          "tag": "v2",
+          "start": null,
+          "end": null,
+          "sources": [
+            {
+              "type": "PARTICIPANT"
+            },
+            {
+              "type": "ACTIVITY",
+              "activityCode": "RELEASE_MINOR",
+              "versionTag": "v2"
+            },
+            {
+              "type": "ACTIVITY",
+              "activityCode": "PARENTAL_CONSENT",
+              "versionTag": "v2"
+            }
+          ],
+          "files": [
+            {
+              "filepath": "pdfs/release_parental_contact_v2.pdf",
+              "type": "CUSTOM",
+              "substitutions": []
+            },
+            {
+              "filepath": "pdfs/release_parental_physician_v2.pdf",
+              "type": "PHYSICIAN",
+              "fields": {
+                "name": "physician_name",
+                "institution": "physician_institution",
+                "city": "physician_city",
+                "state": "physician_state"
+              }
+            },
+            {
+              "filepath": "pdfs/release_parental_agreement_v2.pdf",
+              "type": "CUSTOM",
+              "substitutions": [
+                {
+                  "type": "TEXT",
+                  "field": "first_name",
+                  "activityCode": "PARENTAL_CONSENT",
+                  "questionStableId": "PARENTAL_CONSENT_FIRSTNAME"
+                },
+                {
+                  "type": "TEXT",
+                  "field": "second_name",
+                  "activityCode": "PARENTAL_CONSENT",
+                  "questionStableId": "PARENTAL_CONSENT_LASTNAME"
+                },
+                {
+                  "type": "ACTIVITY_DATE",
+                  "field": "date",
+                  "activityCode": "RELEASE_MINOR"
+                }
+              ]
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "name": "osproject-release-consent-assent",
+      "filename": "osproject-release-consent-assent",
+      "displayName": "Osteo parental consent-assent release pdf",
+      "versions": [
+        {
+          "tag": "v2",
+          "start": null,
+          "end": null,
+          "sources": [
+            {
+              "type": "PARTICIPANT"
+            },
+            {
+              "type": "ACTIVITY",
+              "activityCode": "RELEASE_MINOR",
+              "versionTag": "v2"
+            },
+            {
+              "type": "ACTIVITY",
+              "activityCode": "CONSENT_ASSENT",
+              "versionTag": "v2"
+            }
+          ],
+          "files": [
+            {
+              "filepath": "pdfs/release_parental_contact_v2.pdf",
+              "type": "CUSTOM",
+              "substitutions": []
+            },
+            {
+              "filepath": "pdfs/release_parental_physician_v2.pdf",
+              "type": "PHYSICIAN",
+              "fields": {
+                "name": "physician_name",
+                "institution": "physician_institution",
+                "city": "physician_city",
+                "state": "physician_state"
+              }
+            },
+            {
+              "filepath": "pdfs/release_parental_agreement_v2.pdf",
+              "type": "CUSTOM",
+              "substitutions": [
+                {
+                  "type": "TEXT",
+                  "field": "signature",
+                  "activityCode": "CONSENT_ASSENT",
+                  "questionStableId": "CONSENT_ASSENT_PARENT_SIGNATURE"
+                },
+                {
+                  "type": "ACTIVITY_DATE",
+                  "field": "date",
+                  "activityCode": "RELEASE_MINOR"
+                }
+              ]
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
