--- pepper-apis/studybuilder-cli/studies/osteo/patches/pdf-fixes.conf
@@ -1,74 +1,0 @@
-{
-  "pdfTemplates": [
-    {
-      "name": "osproject-release",
-      "filepath": "pdfs/release_agreement.pdf",
-      "substitutions": [
-        {
-          "type": "TEXT",
-          "field": "first_name",
-          "activityCode": "CONSENT",
-          "questionStableId": "CONSENT_FIRSTNAME"
-        },
-        {
-          "type": "TEXT",
-          "field": "second_name",
-          "activityCode": "CONSENT",
-          "questionStableId": "CONSENT_LASTNAME"
-        },
-        {
-          "type": "ACTIVITY_DATE",
-          "field": "date",
-          "activityCode": "RELEASE_SELF"
-        }
-      ]
-    },
-    {
-      "name": "osproject-release-parental",
-      "filepath": "pdfs/release_parental_agreement.pdf",
-      "substitutions": [
-        {
-          "type": "TEXT",
-          "field": "first_name",
-          "activityCode": "PARENTAL_CONSENT",
-          "questionStableId": "PARENTAL_CONSENT_FIRSTNAME"
-        },
-        {
-          "type": "TEXT",
-          "field": "second_name",
-          "activityCode": "PARENTAL_CONSENT",
-          "questionStableId": "PARENTAL_CONSENT_LASTNAME"
-        },
-        {
-          "type": "ACTIVITY_DATE",
-          "field": "date",
-          "activityCode": "RELEASE_MINOR"
-        }
-      ]
-    },
-    {
-      "name": "osproject-release-consent-assent",
-      "filepath": "pdfs/release_parental_agreement.pdf",
-      "substitutions": [
-        {
-          "type": "TEXT",
-          "field": "first_name",
-          "activityCode": "CONSENT_ASSENT",
-          "questionStableId": "CONSENT_ASSENT_FIRSTNAME"
-        },
-        {
-          "type": "TEXT",
-          "field": "second_name",
-          "activityCode": "CONSENT_ASSENT",
-          "questionStableId": "CONSENT_ASSENT_LASTNAME"
-        },
-        {
-          "type": "ACTIVITY_DATE",
-          "field": "date",
-          "activityCode": "RELEASE_MINOR"
-        }
-      ]
-    },
-  ]
-}
-
